+++ Leomard/Models/Leomard/SessionInfo.swift
@@ -7,11 +7,13 @@
 
 import Foundation
 
-struct SessionInfo: Codable {
+struct SessionInfo: Codable, Equatable, Hashable, Identifiable {
+    var id = UUID()
     public let loginResponse: LoginResponse
     public let lemmyInstance: String
+    public let name: String
     
-    init(loginResponse: LoginResponse, lemmyInstance: String) {
+    init(loginResponse: LoginResponse, lemmyInstance: String, name: String) {
         self.loginResponse = loginResponse
         
         // Remove "http://" and "https://" from lemmyInstance.
@@ -19,5 +21,17 @@ struct SessionInfo: Codable {
         correctedLink = correctedLink.replacingOccurrences(of: "http://", with: "")
         
         self.lemmyInstance = correctedLink
+        self.name = name
+    }
+    
+    static func == (lhs: SessionInfo, rhs: SessionInfo) -> Bool {
+        return lhs.name == rhs.name
+        && lhs.lemmyInstance == rhs.lemmyInstance
+        && lhs.loginResponse.jwt == rhs.loginResponse.jwt
+    }
+    
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(name)
+        hasher.combine(lemmyInstance)
     }
 }